backend/pyproject.toml
[project]
name = "cleark-pyton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "clerk-backend-api>=4.2.0",
    "fastapi>=0.128.4",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
